+++ .babelrc
@@ -8,6 +8,7 @@
     "@babel/plugin-transform-runtime",
     "@babel/plugin-syntax-dynamic-import",
     "@babel/plugin-transform-react-jsx",
-    "babel-plugin-styled-components"
+    "babel-plugin-styled-components",
+    "react-hot-loader/babel"
   ]
 }
