+++ client/src/redux/store.js
@@ -1,8 +1,12 @@
 import { createStore, compose, applyMiddleware } from "redux";
 import thunk from "redux-thunk";
 import { merge } from "./merge";
+import { Provider } from "react-redux";
 import root from "window-or-global";
+import devToolsEnhancer from "remote-redux-devtools";
+import { hydrate } from "react-dom";
 
-const composeEnhancers = root.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
-const enhancer = composeEnhancers(applyMiddleware(thunk));
-export const store = createStore(merge, enhancer);
+const preloadedState = root.__PRELOADED_STATE__;
+delete root.__PRELOADED_STATE__;
+
+export const store = createStore(merge, preloadedState, devToolsEnhancer());
