+++ package.json
@@ -18,6 +18,7 @@
     "express": "^4.17.1",
     "react": "^16.9.0",
     "react-dom": "^16.9.0",
+    "react-hot-loader": "^4.12.14",
     "react-redux": "^7.1.1",
     "redux": "^4.0.4",
     "redux-thunk": "^2.3.0",
@@ -44,6 +45,7 @@
     "html-webpack-plugin": "^3.2.0",
     "prettier": "^1.18.2",
     "react-router-dom": "^5.1.2",
+    "remote-redux-devtools": "^0.5.16",
     "webpack": "^4.40.0",
     "webpack-cli": "^3.3.8",
     "webpack-dev-server": "^3.8.0",
